requirements.txt
asgiref==3.9.1
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
redis==6.4.0
django-redis==6.0.0

